Consolidate glue connection constants
