+++ athena-db2/src/main/java/com/amazonaws/athena/connectors/db2/Db2EnvironmentProperties.java
@@ -23,6 +23,8 @@ import com.amazonaws.athena.connectors.jdbc.JdbcEnvironmentProperties;
 
 import java.util.Map;
 
+import static com.amazonaws.athena.connector.lambda.connection.EnvironmentConstants.DATABASE;
+
 public class Db2EnvironmentProperties extends JdbcEnvironmentProperties
 {
     @Override
