+++ athena-sqlserver/src/main/java/com/amazonaws/athena/connectors/sqlserver/SqlServerEnvironmentProperties.java
@@ -23,6 +23,8 @@ import com.amazonaws.athena.connectors.jdbc.JdbcEnvironmentProperties;
 
 import java.util.Map;
 
+import static com.amazonaws.athena.connector.lambda.connection.EnvironmentConstants.DATABASE;
+
 public class SqlServerEnvironmentProperties extends JdbcEnvironmentProperties
 {
     @Override
