+++ athena-synapse/src/main/java/com/amazonaws/athena/connectors/synapse/SynapseEnvironmentProperties.java
@@ -23,6 +23,8 @@ import com.amazonaws.athena.connectors.jdbc.JdbcEnvironmentProperties;
 
 import java.util.Map;
 
+import static com.amazonaws.athena.connector.lambda.connection.EnvironmentConstants.DATABASE;
+
 public class SynapseEnvironmentProperties extends JdbcEnvironmentProperties
 {
     @Override
